Provide `ahkpm init --defaults` flag to create ahkpm.json with defaults and no prompts
Fixes #134
